+++ src/features/profile/components/index.js
@@ -1,3 +1,4 @@
 export { Posts } from "./Posts";
 export { Followers } from "./Followers";
 export { Following } from "./Following";
+export { EditProfileForm } from "./EditProfileForm";
